move StringEV
